// lib/plugins/generated/Label_1L_Slash.ts
// AUTO-GENERATED from /Users/kevin/Cloud/Dropbox/work/airframes/acars-decoder-typescript/vendor/airframes-decoder/spec/labels/1L/Slash.yaml. Do not edit.
// Plugin: Label_1L_Slash

import { DecoderPlugin } from "@airframes/ads-runtime-ts";
import type { DecodeResult, Message, Options } from "@airframes/ads-runtime-ts";
import { ResultFormatter } from "@airframes/ads-runtime-ts";
import * as helpers from "@airframes/ads-runtime-ts/helpers";
import * as hatches from "../escape_hatches";

export class Label_1L_Slash extends DecoderPlugin {
  name = "label-1l-slash";

  qualifiers() {
    return {
      labels: ["1L"],
      preambles: ["+","-"],
    };
  }

  decode(message: Message, options: Options = {}): DecodeResult {
    const result = this.initResult(message, "Position Report");

    return hatches.label_1l_slash_decode(this, message, result, options);
  }
}